extract error types into errors.go
